+++ src/main/java/model/card/CardType.java
@@ -1,4 +1,4 @@
-package model;
+package model.card;
 
 public enum CardType {
