+++ src/main/java/model/card/Cards.java
@@ -1,4 +1,4 @@
-package model;
+package model.card;
 
 import java.util.ArrayList;
 import java.util.HashSet;
