+++ src/main/java/model/deck/DeckCreateStrategy.java
@@ -1,6 +1,7 @@
-package model;
+package model.deck;
 
 import java.util.List;
+import model.card.Card;
 
 public interface DeckCreateStrategy {
