+++ src/main/java/model/deck/DeckStrategy.java
@@ -1,4 +1,6 @@
-package model;
+package model.deck;
+
+import model.card.Card;
 
 public interface DeckStrategy {
