+++ src/test/java/model/CardTest.java
@@ -2,6 +2,9 @@ package model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import model.card.Card;
+import model.card.CardNumber;
+import model.card.CardType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
