+++ src/test/java/model/CardsTest.java
@@ -1,7 +1,9 @@
-import model.Card;
-import model.CardNumber;
-import model.CardType;
-import model.Cards;
+package model;
+
+import model.card.Card;
+import model.card.CardNumber;
+import model.card.CardType;
+import model.card.Cards;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
