+++ src/test/java/model/DeckTest.java
@@ -1,6 +1,9 @@
 package model;
 
 import java.util.List;
+import model.card.Card;
+import model.deck.Deck;
+import model.deck.DeckCreateStrategy;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
